feat(move): New place for home components
