+++ src/components/Home.js
@@ -1,9 +1,13 @@
 import React from 'react';
-import HeroSection from './HeroSection';
+import HeroSection from '../components/Home/HeroSection';
+import WhatIsSepsis from '../components/Home/WhatIsSepsis';
 
 const Home = () => {
   return (
+    <div className="home">
       <HeroSection />
+      <WhatIsSepsis />
+    </div>
   );
 };
